requirements.txt
python-dotenv>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.2.0
mcp[cli]>=1.0.0
